requirements.txt
camelot-py[cv]
pandas
openpyxl
numpy
opencv-python
ghostscript
pdfplumber
